% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherCombPval.R
\name{FisherCombPval}
\alias{FisherCombPval}
\title{Calulate p-values using  Fisher combination statistic}
\usage{
FisherCombPval(P.values, Pheno, kendall.only = FALSE, miss.pval = FALSE,
  verbose = FALSE)
}
\arguments{
\item{P.values}{A matrix or data frame for p-values. Each column represents marginal p-values.}

\item{Pheno}{A matrix or data frame for the observed outcome. Each column represent the marginal phenotype.}

\item{kendall.only}{A logical variable indicates whether to use the Kindall rank correlation statistic to estimate the correlation matrix from the multivaraite phenotypes or estimate it based on types (continuous, binary, or ordinal) of multivarite phenotypes (default)}

\item{miss.pval}{A logical variable indicates whether the marginal p-values contain missing values.}

\item{verbose}{A logical variable indicates whether to show the details of analysis process}
}
\description{
This funciton calculate the omnibus p-values from the marginal p-values and outcome matrix
}
\examples{
set.seed(1)
pval <- matrix(runif(10*6),10,6)
pheno <- data.frame(y1=rnorm(20),
                    y2=rnorm(20),
                    y3=rnorm(20) > 0,
                    y4=sample(c("M","F"),20,replace=TRUE),
                    y5=ordered(cut(rnorm(20),4)),
                    y6=ordered(cut(rnorm(20),5))
                   )
FisherCombPval(P.values=pval,
              Pheno=pheno,
              kendall.only=FALSE,
              miss.pval=FALSE)
}
